Extract SearchLogProcessor module
