+++ lib/log-parsers.js
@@ -1,15 +1,5 @@
 'use strict'
 
-const jmespath = require('jmespath')
-
-function search (input, { search }) {
-  if (search && !jmespath.search(input, search)) {
-    return { output: undefined, done: true }
-  } else {
-    return { output: input }
-  }
-}
-
 function ignore (input, { ignore }) {
   const ignoreKeys = ignore ? new Set(ignore.split(',')) : undefined
   if (ignoreKeys) {
@@ -26,7 +16,6 @@ function ignore (input, { ignore }) {
 }
 
 const builtInLogParsers = {
-  search,
   ignore
 }
