+++ lib/log-processors.js
@@ -4,8 +4,9 @@ const { builtInLineBuilders } = require('./line-builders')
 const LogProcessor = require('./LogProcessor')
 const JsonLogProcessor = require('./processors/JsonLogProcessor')
 const PrimitivesLogProcessor = require('./processors/PrimitivesLogProcessor')
+const SearchLogProcessor = require('./processors/SearchLogProcessor')
 
-const { search, ignore } = builtInLogParsers
+const { ignore } = builtInLogParsers
 const {
   appendLevel,
   appendMessage,
@@ -28,12 +29,6 @@ function createLogProcessor (definition) {
     return new builtInLogProcessors[definition]()
   } else {
     return definition
-  }
-}
-
-class SearchLogProcessor extends LogProcessor {
-  parse (input, context) {
-    return search(input, context)
   }
 }
